UCM-14: extend httpx version, split update for Python > 3.6 and below
