+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "pyuploadcare"
-version = "3.0.0"
+version = "3.1.0"
 description = "Python library for Uploadcare.com"
 authors = ["Uploadcare Inc <hello@uploadcare.com>"]
 readme = "README.rst"
@@ -26,7 +26,10 @@ ucare = 'pyuploadcare.ucare_cli.main:main'
 
 [tool.poetry.dependencies]
 python = "^3.6.2"
-httpx = "^0.18.2"
+httpx = [
+  {version = "^0.18.2", python = ">=3.6,<3.7"},
+  {version = "^0.23.0", python = "^3.7"}
+]
 pydantic = {extras = ["email"], version = "^1.8.2"}
 python-dateutil = "^2.8.2"
 pytz = "^2021.1"
@@ -42,7 +45,7 @@ pytest = [
     {version = "^7.1", python = "^3.7"}
 ]
 tox = "^3.24.1"
-black = "^21.7b0"
+black = "^22.3.0"
 isort = "^5.9.3"
 flake8 = "^3.9.2"
 mypy = "^0.910"
